amsvis/amsnet: Replace pickle with json.
Due to security concerns, data to be sent over the network
will be serialized via json instead of pickle.  The socket file
interface will also be removed as it did not play nice with the
json load method.  Data transfers will be implemented with the
socket send, sendall, and recv methods.  These implementations along with
json serialization are contained within the functions send_json_message
and receive_json_message, which allow the client and server to transfer
json messages of variable size.

Signed-off-by: Thomas Falcon <tlfalcon@linux.vnet.ibm.com>